+++ examples/client/config.hpp
@@ -162,6 +162,7 @@ struct InputInterface {
 
 struct InputConfig {
     std::vector<InputInterface> inputs;
+    bool disable_pipe_buffer_optimization;
 };
 
 struct AssistanceDataConfig {